+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - \#10 - When the collection is empty, the search client throws exceptions
 - \#15 - Search won't return status when return nq = 0
 - \#20 - Receive large dataset failed
+- \#21 - go sdk can not return error message if grpc failed
 
 ### Improvement
 ---